+++ packages/domains/src/services/api/index.ts
@@ -1,4 +1,6 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { AzionClientOptions, AzionDomain } from '../../types';
+import { fetchWithErrorHandling } from '../../utils/index';
 import {
   ApiAzionDomainResponse,
   ApiAzionDomainResult,
@@ -9,7 +11,6 @@ import {
 const BASE_URL =
   process.env.AZION_ENV === 'stage' ? 'https://stage-api.azion.net/domains' : 'https://api.azionapi.net/domains';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 const handleApiError = (fields: string[], data: any) => {
   let error = undefined;
   fields.forEach((field: string) => {
@@ -56,23 +57,26 @@ const createDomain = async (
         crl_list: domain.mtls.crlList,
       };
     }
-
-    const response = await fetch(BASE_URL, {
-      method: 'POST',
-      headers: makeHeaders(token),
-      body: JSON.stringify(body),
-    });
-    const data = await response.json();
-    if (debug) console.log('Response Post Domains', data);
+    const data = await fetchWithErrorHandling(
+      BASE_URL,
+      {
+        method: 'POST',
+        headers: makeHeaders(token),
+        body: JSON.stringify(body),
+      },
+      debug,
+    );
     if (!data.results) {
       return {
         error: { message: handleApiError(['detail', 'edge_application_id'], data), operation: 'create domain' },
       };
     }
     return data;
-  } catch (error) {
+  } catch (error: any) {
     if (debug) console.error('Error creating Domain:', error);
-    throw error;
+    return {
+      error: { message: error.toString(), operation: 'create domain' },
+    };
   }
 };
 
@@ -84,12 +88,14 @@ const getDomains = async (
   try {
     const { page_size = 10, page = 1 } = queryParams || {};
     const queryParamsUrl = new URLSearchParams({ page_size: String(page_size), page: String(page) });
-    const response = await fetch(`${BASE_URL}?${queryParamsUrl.toString()}`, {
-      method: 'GET',
-      headers: makeHeaders(token),
-    });
-    const data = await response.json();
-    if (options?.debug) console.log('Response List Domains', data);
+    const data = await fetchWithErrorHandling(
+      `${BASE_URL}?${queryParamsUrl.toString()}`,
+      {
+        method: 'GET',
+        headers: makeHeaders(token),
+      },
+      options?.debug,
+    );
     if (!data.results) {
       return {
         error: { message: handleApiError(['detail', 'edge_application_id'], data), operation: 'list domains' },
@@ -117,9 +123,11 @@ const getDomains = async (
           : undefined,
       })),
     };
-  } catch (error) {
+  } catch (error: any) {
     if (options?.debug) console.error('Error listing Domains:', error);
-    throw error;
+    return {
+      error: { message: error.toString(), operation: 'list domains' },
+    };
   }
 };
 
@@ -129,21 +137,25 @@ const getDomainById = async (
   options?: AzionClientOptions,
 ): Promise<ApiAzionDomainResult> => {
   try {
-    const response = await fetch(`${BASE_URL}/${domainId}`, {
-      method: 'GET',
-      headers: makeHeaders(token),
-    });
-    const data = await response.json();
-    if (options?.debug) console.log('Response Get Domain', data);
+    const data = await fetchWithErrorHandling(
+      `${BASE_URL}/${domainId}`,
+      {
+        method: 'GET',
+        headers: makeHeaders(token),
+      },
+      options?.debug,
+    );
     if (!data.results) {
       return {
         error: { message: handleApiError(['detail', 'edge_application_id'], data), operation: 'get domain' },
       };
     }
     return data;
-  } catch (error) {
+  } catch (error: any) {
     if (options?.debug) console.error('Error getting Domain:', error);
-    throw error;
+    return {
+      error: { message: error.toString(), operation: 'get domain' },
+    };
   }
 };
 
@@ -178,22 +190,26 @@ const updateDomain = async (
       };
     }
 
-    const response = await fetch(`${BASE_URL}/${domainId}`, {
-      method: 'PUT',
-      headers: makeHeaders(token),
-      body: JSON.stringify(body),
-    });
-    const data = await response.json();
-    if (options?.debug) console.log('Response Put Domains', data);
+    const data = await fetchWithErrorHandling(
+      `${BASE_URL}/${domainId}`,
+      {
+        method: 'PUT',
+        headers: makeHeaders(token),
+        body: JSON.stringify(body),
+      },
+      options?.debug,
+    );
     if (!data.results) {
       return {
         error: { message: handleApiError(['detail', 'edge_application_id'], data), operation: 'update domain' },
       };
     }
     return data;
-  } catch (error) {
+  } catch (error: any) {
     if (options?.debug) console.error('Error updating Domain:', error);
-    throw error;
+    return {
+      error: { message: error.toString(), operation: 'update domain' },
+    };
   }
 };
 
@@ -203,23 +219,21 @@ const deleteDomain = async (
   options?: AzionClientOptions,
 ): Promise<ApiAzionDomainResult> => {
   try {
-    const result = await fetch(`${BASE_URL}/${domainId}`, {
-      method: 'DELETE',
-      headers: makeHeaders(token),
-    });
-
-    if (result.status !== 204) {
-      const message = await result.json();
-      return {
-        error: { message: handleApiError(['detail', 'edge_application_id'], message), operation: 'delete domain' },
-      };
-    }
-
-    if (options?.debug) console.log('Response Delete Domain');
+    await fetchWithErrorHandling(
+      `${BASE_URL}/${domainId}`,
+      {
+        method: 'DELETE',
+        headers: makeHeaders(token),
+      },
+      options?.debug,
+      false,
+    );
     return { results: undefined };
-  } catch (error) {
+  } catch (error: any) {
     if (options?.debug) console.error('Error deleting Domain:', error);
-    throw error;
+    return {
+      error: { message: error.toString(), operation: 'delete domain' },
+    };
   }
 };